+++ components/VideoUploadForm.js
@@ -1,10 +1,41 @@
+import { Box, Flex, Text, Button } from "rebass";
+import { Label, Input, Select, Textarea, Radio, Checkbox } from "@rebass/forms";
+
+import { base } from "@theme-ui/presets";
+import { jsx } from "theme-ui";
+
 import { instanceOf } from "prop-types";
 import cookie from "react-cookies";
 import Router from "next/router";
 
 const axios = require("axios").default;
 
-class VideoUploadForm extends React.Component {
+const videoUploadWrap = {
+  border: "2px dashed blue",
+  borderRadius: "5px",
+  position: "relative"
+};
+
+const videoUploadText = {
+  textAlign: "center",
+  fontWeight: "100",
+  textTransform: "uppercase",
+  color: "black",
+  padding: "7px 0"
+};
+
+const videoUploadInput = {
+  position: "absolute",
+  margin: "0",
+  padding: "0",
+  width: "100%",
+  height: "100%",
+  outline: "none",
+  opacity: "0",
+  cursor: "pointer"
+};
+
+class VideoUploadFormTemp extends React.Component {
   constructor(props) {
     super(props);
   }
@@ -44,7 +75,6 @@ class VideoUploadForm extends React.Component {
             httpOnly: true
           }
         );
-
         Router.push("/layout", "/layout");
       })
       .catch(function(error) {
@@ -54,25 +84,45 @@ class VideoUploadForm extends React.Component {
 
   render() {
     return (
-      <div>
-        <form
-          action="/upload"
-          method="post"
-          encType="multipart/form-data"
-          target="transFrame"
-        >
-          <input type="file" name="video" onChange={this.onChangeHandler} />
-          <button onClick={this.onClickHandler}>upload</button>
-        </form>
-        <iframe
-          width="200"
-          height="100"
-          name="transFrame"
-          id="transFrame"
-        ></iframe>
-      </div>
+      <Box
+        as="form"
+        onSubmit={e => e.preventDefault()}
+        py={3}
+        action="/upload"
+        method="post"
+        encType="multipart/form-data"
+      >
+        <Flex>
+          <Box style={videoUploadWrap} width={4 / 10} mx={2}>
+            <Input
+              style={videoUploadInput}
+              type="file"
+              name="video"
+              onChange={this.onChangeHandler}
+            />
+            <Text style={videoUploadText}>Upload Video</Text>
+          </Box>
+          <Input
+            id="deviceType"
+            name="deviceType"
+            placeholder="deviceType"
+            width={4 / 10}
+            mx={2}
+          />
+          <Button
+            width={2 / 10}
+            mx={2}
+            type="submit"
+            id="submit"
+            name="submit"
+            onClick={this.onClickHandler}
+          >
+            upload
+          </Button>
+        </Flex>
+      </Box>
     );
   }
 }
 
-export default VideoUploadForm;
+export default VideoUploadFormTemp;
